+++ expected/create.out
@@ -1,6 +1,6 @@
 CREATE SCHEMA db_create;
 CREATE EXTENSION influx WITH SCHEMA db_create;
-SELECT pg_sleep(1) FROM db_create.worker_launch(4711::text);
+SELECT pg_sleep(1) FROM db_create.launch_udp_worker(4711::text);
  pg_sleep 
 ----------
  
@@ -8,7 +8,7 @@ SELECT pg_sleep(1) FROM db_create.worker_launch(4711::text);
 
 -- Should create the table
 \d db_create.*
-CALL db_create.send_packet('system,host=fury uptime=607641i 1574753954000000000', 4711::text);
+CALL db_create.send_udp_packet('system,host=fury uptime=607641i 1574753954000000000', 4711::text);
 SELECT pg_sleep(1);
  pg_sleep 
 ----------
@@ -33,7 +33,7 @@ BEGIN
 END;
 $$ LANGUAGE plpgsql;
 \d db_create.*
-CALL db_create.send_packet('disk,device=nvme0n1p1 free=527806464i 1574753954000000000', 4711::text);
+CALL db_create.send_udp_packet('disk,device=nvme0n1p1 free=527806464i 1574753954000000000', 4711::text);
 SELECT pg_sleep(1);
  pg_sleep 
 ----------
@@ -52,7 +52,7 @@ DROP TABLE db_create.disk;
 ALTER EXTENSION influx DROP FUNCTION db_create._create;
 DROP FUNCTION db_create._create;
 \d db_create.*
-CALL db_create.send_packet('cpu,host=fury uptime=607641i 1574753954000000000', 4711::text);
+CALL db_create.send_udp_packet('cpu,host=fury uptime=607641i 1574753954000000000', 4711::text);
 SELECT pg_sleep(1);
  pg_sleep 
 ----------
