+++ expected/inval.out
@@ -3,11 +3,11 @@ CREATE TABLE db_worker.cpu(_time timestamptz, _tags jsonb, _fields jsonb);
 CREATE EXTENSION influx WITH SCHEMA db_worker;
 \set VERBOSITY terse
 \x on
-SELECT pg_sleep(1) FROM db_worker.worker_launch('db_worker', 4711::text);
+SELECT pg_sleep(1) FROM db_worker.launch_udp_worker('db_worker', 4711::text);
 -[ RECORD 1 ]
 pg_sleep | 
 
-CALL db_worker.send_packet('cpu,cpu=cpu0,host=fury usage_system=2.04,usage_user=5.40 1574753954000000000', 4711::text);
+CALL db_worker.send_udp_packet('cpu,cpu=cpu0,host=fury usage_system=2.04,usage_user=5.40 1574753954000000000', 4711::text);
 SELECT pg_sleep(1);
 -[ RECORD 1 ]
 pg_sleep | 
@@ -19,7 +19,7 @@ _tags   | {"cpu": "cpu0", "host": "fury"}
 _fields | {"usage_user": "5.40", "usage_system": "2.04"}
 
 ALTER TABLE db_worker.cpu ADD COLUMN cpu text;
-CALL db_worker.send_packet('cpu,cpu=cpu0,host=fury usage_system=2.04,usage_user=5.40 1574753955000000000', 4711::text);
+CALL db_worker.send_udp_packet('cpu,cpu=cpu0,host=fury usage_system=2.04,usage_user=5.40 1574753955000000000', 4711::text);
 SELECT pg_sleep(1);
 -[ RECORD 1 ]
 pg_sleep | 
@@ -37,7 +37,7 @@ _fields | {"usage_user": "5.40", "usage_system": "2.04"}
 cpu     | cpu0
 
 ALTER TABLE db_worker.cpu ADD COLUMN host text;
-CALL db_worker.send_packet('cpu,cpu=cpu0,host=fury usage_system=2.04,usage_user=5.40 1574753956000000000', 4711::text);
+CALL db_worker.send_udp_packet('cpu,cpu=cpu0,host=fury usage_system=2.04,usage_user=5.40 1574753956000000000', 4711::text);
 SELECT pg_sleep(1);
 -[ RECORD 1 ]
 pg_sleep |
