+++ influx.c
@@ -34,7 +34,9 @@
 #include <stdbool.h>
 #include <string.h>
 
+#include "guc.h"
 #include "ingest.h"
+#include "udp.h"
 #include "worker.h"
 
 PG_MODULE_MAGIC;
@@ -42,21 +44,6 @@ PG_MODULE_MAGIC;
 PG_FUNCTION_INFO_V1(parse_influx);
 
 void PGDLLEXPORT _PG_init(void);
-
-/** Number of Influx protocol workers. */
-static int InfluxWorkersCount;
-
-/** Service name to listen on. */
-static char *InfluxServiceName;
-
-/** Schema name to use for metrics. */
-static char *InfluxSchemaName;
-
-/** Database name to work with. */
-static char *InfluxDatabaseName;
-
-/** Role name to use when connecting to the database. */
-static char *InfluxRoleName;
 
 /** Parser state setup. */
 IngestState *ParseInfluxSetup(char *buffer) {
@@ -147,11 +134,9 @@ Datum parse_influx(PG_FUNCTION_ARGS) {
 static void StartBackgroundWorkers(const char *database_name,
                                    const char *schema_name,
                                    const char *role_name,
-                                   const char *service_name, int worker_count) {
+                                   const char *service_name) {
   MemoryContext oldcontext = MemoryContextSwitchTo(TopMemoryContext);
   WorkerArgs args = {0};
-  int i;
-  BackgroundWorker worker;
 
   if (schema_name)
     strncpy(args.namespace, schema_name, sizeof(args.namespace));
@@ -162,63 +147,17 @@ static void StartBackgroundWorkers(const char *database_name,
   if (service_name)
     strncpy(args.service, service_name, sizeof(args.service));
 
-  elog(LOG, "starting influx workers");
-
-  InfluxWorkerInit(&worker, &args);
-  elog(LOG, "memory allocated in %s memory context",
-       CurrentMemoryContext->name);
-  for (i = 0; i < worker_count; i++)
-    RegisterBackgroundWorker(&worker);
-  elog(LOG, "background workers started");
+  StartUdpBackgroundWorkers(&args);
   MemoryContextSwitchTo(oldcontext);
 }
 
 void _PG_init(void) {
-  DefineCustomIntVariable("influx.workers",     /* option name */
-                          "Number of workers.", /* short descriptor */
-                          "Number of workers to spawn when starting up"
-                          "the server.",       /* long description */
-                          &InfluxWorkersCount, /* value address */
-                          4,                   /* boot value */
-                          0,                   /* min value */
-                          20,                  /* max value */
-                          PGC_SIGHUP,          /* option context */
-                          0,                   /* option flags */
-                          NULL,                /* check hook */
-                          NULL,                /* assign hook */
-                          NULL);               /* show hook */
+  InfluxGucInitDynamic();
 
   if (!process_shared_preload_libraries_in_progress)
     return;
 
-  /* Right now we have only UDP supported, so we use 8089. If we decide to
-   * support more protocols, we should dynamically pick the default based on the
-   * protocol. */
-  DefineCustomStringVariable(
-      "influx.service", "Service name.",
-      "Service name to listen on, or port number. If it is a service name, it"
-      " will be looked up.If it is a port, it will be used as it is.",
-      &InfluxServiceName, "8089", PGC_POSTMASTER, 0, NULL, NULL, NULL);
-  DefineCustomStringVariable(
-      "influx.database", "Database name.", "Database name to connect to.",
-      &InfluxDatabaseName, NULL, PGC_POSTMASTER, 0, NULL, NULL, NULL);
-  DefineCustomStringVariable(
-      "influx.role", "Role name.",
-      "Role name to use when connecting to the database. Default is to"
-      " connect as superuser.",
-      &InfluxRoleName, NULL, PGC_POSTMASTER, 0, NULL, NULL, NULL);
-  DefineCustomStringVariable(
-      "influx.schema", "Schema name.",
-      "Schema name to use for the workers. This is where the measurement"
-      " tables should be placed.",
-      &InfluxSchemaName, NULL, PGC_POSTMASTER, 0, NULL, NULL, NULL);
-
-  elog(LOG,
-       "InfluxDatabaseName: %s, InfluxSchemaName: %s, InfluxServiceName: %s, "
-       "InfluxRoleName: %s, InfluxWorkersCount: %d",
-       InfluxDatabaseName, InfluxSchemaName, InfluxServiceName, InfluxRoleName,
-       InfluxWorkersCount);
-
+  InfluxGucInitStatic();
   StartBackgroundWorkers(InfluxDatabaseName, InfluxSchemaName, InfluxRoleName,
-                         InfluxServiceName, InfluxWorkersCount);
+                         InfluxServiceName);
 }
