+++ sql/create.sql
@@ -1,10 +1,10 @@
 CREATE SCHEMA db_create;
 CREATE EXTENSION influx WITH SCHEMA db_create;
-SELECT pg_sleep(1) FROM db_create.worker_launch(4711::text);
+SELECT pg_sleep(1) FROM db_create.launch_udp_worker(4711::text);
 
 -- Should create the table
 \d db_create.*
-CALL db_create.send_packet('system,host=fury uptime=607641i 1574753954000000000', 4711::text);
+CALL db_create.send_udp_packet('system,host=fury uptime=607641i 1574753954000000000', 4711::text);
 SELECT pg_sleep(1);
 \d db_create.*
 DROP TABLE db_create.system;
@@ -19,7 +19,7 @@ END;
 $$ LANGUAGE plpgsql;
 
 \d db_create.*
-CALL db_create.send_packet('disk,device=nvme0n1p1 free=527806464i 1574753954000000000', 4711::text);
+CALL db_create.send_udp_packet('disk,device=nvme0n1p1 free=527806464i 1574753954000000000', 4711::text);
 SELECT pg_sleep(1);
 \d db_create.*
 DROP TABLE db_create.disk;
@@ -29,7 +29,7 @@ ALTER EXTENSION influx DROP FUNCTION db_create._create;
 DROP FUNCTION db_create._create;
 
 \d db_create.*
-CALL db_create.send_packet('cpu,host=fury uptime=607641i 1574753954000000000', 4711::text);
+CALL db_create.send_udp_packet('cpu,host=fury uptime=607641i 1574753954000000000', 4711::text);
 SELECT pg_sleep(1);
 \d db_create.*
